pyproject.toml
[project]
name = "atelie-catalog-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "fastapi[standard]>=0.136.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
    "taskipy>=1.14.1",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations', 'alembic']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']


[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
# Linters e Formatação 
lint = 'ruff check . && ruff check --diff .'
format = 'ruff check . --fix && ruff format .'

# Testes
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vv'
post_test = 'coverage html'

# Execução
run = 'fastapi dev app/main.py'
